+++ app/controllers/answers_controller.rb
@@ -18,13 +18,7 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    if current_user.author_of?(@answer)
-      @answer.destroy
-      flash[:notice] = 'Your answer successfully deleted.'
-    else
-      flash[:alert] = 'You have not rights to delete this answer!'
-    end
-    redirect_to @answer.question
+    @answer.destroy if current_user.author_of?(@answer)
   end
 
   private
