+++ app/views/answers/destroy.js.erb
@@ -1,0 +1,1 @@
+$('#answer-<%= @answer.id %>').remove();
